客户交付包_v2.0.0/requirements.txt
requests>=2.31.0
pyaudio>=0.2.14
websocket-client>=1.8.0
chardet>=5.0.0
